Ramin | My CV Website | 2024-05-06 21:21 | feature/feature-backend | Feat: Implement References model associated with Info
In this commit, the References model has been implemented to link directly to the Info model, facilitating the storage of reference details pertinent to the user's Info records. Each reference is tied to a specific Info object, enhancing the ability to store and retrieve comprehensive background information on users, including endorsements or references that are critical to professional profiles. Similar to Experiences, the CASCADE delete rule ensures that references are managed in line with the lifecycle of the parent Info object, thereby maintaining data coherence and cleanliness.